align asr benchmark with tts code style.
